Created new ImageUtils class
